Use npx and create-react-app to make a new app.
