+++ README.md
@@ -1,2 +1,19 @@
 # react-lake-app
+
 App from Eve Porcello's course
+
+## How to Create a New React App (With Create-React-App)
+
+1. Open a terminal session to the directory you would like your project to be stored.
+
+```sh
+C:\Users\Alex\Documents>
+```
+
+2. Use NPX to create a new react app using create-react-app in that directory with the name of the project.
+
+```sh
+C:\Users\Alex\Documents>npx create-react-app project-name
+```
+
+You don't need to worry about Webpack or Babel if you use create-react-app :)
