+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "lake-app",
+  "name": "react-lake-app",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
