+++ src/constants.js
@@ -1,0 +1,17 @@
+export const lakeList = [
+  {
+    id: 1,
+    name: "Echo Lake",
+    trailhead: "Echo Lake"
+  },
+  {
+    id: 2,
+    name: "Maud Lake",
+    trailhead: "Wright's Lake"
+  },
+  {
+    id: 3,
+    name: "Cascade Lake",
+    trailhead: "Bayview"
+  }
+];
